First shot at doxygen documentation for minisip.  To generate the documentation, run
	cd Documentation
	doxygen doxygen.conf

The current doxygen.conf will generate documentation for minisip and libmsip.
Some of the information in the devel_guide.txt should be moved to the code
itself.  This way we ensure that the documentation stays in sync.  The
\addtogroup minisip command might be useful to add comments in the code that
we want to appear in the developer guide.

I have only checked in one UML diagram, as it is not yet clear how we are
going to handle those.  Using dot format would be easier on svn (as they are
text format) and on doxygen (as they can be easyly modified and embedded in
the text docs).  I'll check out Autodia ( http://droogs.org/autodia/ ).

Serious formatting is still needed.