Now the GTK GUI closes correctly (i think under most circumstances). This means: - de register (use previous patch to allow expires=0) - hang up calls (uses call_terminated_early new command) - it should hang up also conference calls
It is not a timed out shutdown. That is, if something goes wrong,
for example, proxy not replying, the time it takes to close is equal 
to that needed to detect a transport_error (>30 secs).

This needs to be "backported" to other guis.
 
Also it may be implemented using some other mechanism, such as a 
StateMachine (and also using timers then).