Another bug which was rendering MiniSIP useless. It turns out one index used to access an array of strings in the TimeStamp class was going out of bounds.
I just added a bound check and reset ... I have no idea if that is the 
correct solution, but at least now MiniSIP does not SegFault on me anymore.

I think there should be a way to turn off the timestamp ... to not even compile it? Or
it does not take a lot of resources?