Fixed the mute-all-but-one senders functionality. Now, the muted sessions send 1 zeroed (no data) packet every MAX (for now, 50, which for audio is 1 second).
I will add later an entry to the config file to (de)activate
this feature (default, i think, should be on). What do 
you think?

The mute is done via the mediahandler (session registry). 
The zeroing of muted sessions' data is done in 
AudioMedia (for audio) and Media classes (for the rest, that 
is, video).