Adapt the code to the new SipURI class.
This should be just the begging. In several more classes uris are badly 
dealt with, creating extra code to be added to functions, or strange 
behaviors (heavily unpredictable). 

Hopefully this fixes some errors and leaves minisip in a working state
at least as good as before. I did some testing, but not all messages
(subscribes, notifys, refer, ... all this are untested).