Fix the Instant Messaging in minisip. I tested agains kphone and snom and it works, both ways.
The only glitch i can find is that the matching of IMs
from and to minisip is not too good. For example:
from gui you send message to "cesc"
the messages is correctly appended the domain "minisip.org", but at the sipstack level,
    thus the gui still thinks it is talking to "cesc"
when the other side replies to the message, minisip gui opens a new window,
    as the message TO is for cesc@minisip.org ... but it should go to the still 
    open IM message ... 
I think this could be fixed by creating some kind of SipIMDialog (a fake one, IM
does not create a dialog ... but then all this could be dealt with).
Also with this (as it is done with sipregisterdialog ... now 2xx, 3xx, 4xx ... they
are never notified to the gui ... thus we are never sure if the message got to the
destination.

Cesc