Fixes and improvements by Mikael Magnusson (it is becoming so usual i don't have to look up the name anymore :D )
the s_l variable needs to be initialize to the RTP sequence number of
the first seen SRTP packet, according to section 3.3.1 of RFC 3711.

I'm resending my patch that solved the problem.

Another patch is also included, that fixes some bugs and removes the
unnecessary rebuild of the whole packet in rtp_authenticate. It's enough
to build the header, the content can be used as is.

Regards,
Mikael