Added a new tag in the config file: <auto_detect_proxy> ... before, this feature was done when <proxy_addr> was empty ... it was a bit not-obvious how it worked, and once minisip had discovered once the proxy and port, it would not autodetect anymore.
Now it autodetects every time it loads.
Also, improved the autodetect mechanism, so it will fetch SRV records according to the
transport selected for the proxy. 
Moreover, if transport is tcp and the lookup fails, it will try udp (it won't if transport
is tls, as udp would be an insecure choice).

Important is the cleanup of the SipDialogConfig file interface (which includes
SipProxy, SipIdentity, ... ). Moved a lot of functionality into these classes, as well
as hid some attributes, now only accessible via functions (see the next commit ... which is mainly
compilation fixes due to this changes).
Now the behavior is the same from the gui and from the load of the config file ... in terms
of autodetect method.

Cesc