This update changes how sip messages are sent by the transport layer, and makes minisip more RFC 3261 compliant and partly compliant to the required server behavior in RFC 3581. If minisip is running on a multi homed system it isn't guaranteed that the following requirement in RFC 3581 holds.
    The response MUST be sent from the same address and port that the
    corresponding request was received on.


Added SipRequest a base class for all classes implementing request
methods, and also used to represent unknown request methods. The
getString function has been removed in all sub classes and is
implemented by SipRequest instead. 

SipRequest also contains addRoute functions which adds a route to the
outbound proxy that should be used for sending the request.

The route header replaces the old way of forwarding all request and
responses (!) to an outbound proxy.

The destination parameters is removed from the public sendMessage
function in SipMessageTransport. They aren't needed since a route
header is used to specify an outbound proxy server for
requests. Responses are sent using the socket the request was received
from, if a stream oriented transport is used, or to the address in the
received parameter of the top via if it was added during reception, or
to the sent-by address otherwise.

One remaining problem is that subsequent requests in a dialog still can
be sent to the wrong outbound proxy, if several sip accounts
using different proxies are defined.