Fix the win32 showstopper ... DNS SRV resolution ...
For linux - use libresolv
For win32-mingw32 - use WinDns
For wince - use getaddrinfo (untested).

The funny thing is that getaddrinfo is available on all platforms ... but in linux it does not support
SRV records, but it does in windows.
Still, in win32-mingw, getaddrinfo is for XP or above ... whereas windns is available for a wider choice ... so we use windns.

For now, the choice of library to use is done in NeworkFunctions.cxx ... via some conditional compilation ... 
USE_WIN32_API_WINDNS ... for windns
USE_WIN32_API_GETADDRINFO ... for wince
ideally, this should be automatically done in the configure.ac ... don't have the time now ...