NEW Feature: Call Recording It allows to record each call separately to a file. It does so on a 2 channel, 8Khz, 16 bit, raw file. Each side of the call (incoming and outgoing audio) is recorded separately to a channel (it is clearer to listen than if everything was mixed).
By hooking a CallRecorder object into the MediaStreamReceiver and the SoundCallBacks.
It uses the FileSoundDevice to write the recorded call to a file.
There is a new button on the GTK gui (a red circle, next to the silence headset and microphone).

BTW, it needs testing and improvement ... I'll do it in the coming days. But it already works ... with some perks :)

Cesc