Adds support for showing modal dialogs in the text ui. It is useful if the SIP stack requires credentials to authenticate to a proxy (at any time ask for user/pass).
This implements support for asking a user questions. You do this by
giving the TextUI a set of questions that the user must answer.

When the dialog has been answered the prompt with any input on it will
be restored.

Note that this needs non-blocking stdin I/O.

(have tried to do the same for GTK, but glade-2 has been giving me a hard
time tonight...)