There are a few issues with how minisip treats configuration data.
For example, the data in SipSoftPhoneConfiguration is global, and used by
ongoing dialogs the same time as the GUI can modify it.

We also treat the configuration data as if it always loaded at start up,
and saved when the user changes it. This makes it a bit hard to implement
things such as signing in and off accounts at any time.

The SipSoftPhoneConfiguration class feels a bit wrong, and with this
commit I start decreasing its importance.

 * The dialog logic only uses two boolean values from
   SipSoftPhoneConfiguration (useSTUN and useAnat). Let's provide
   those values at start up. We then will not have a problem if the user
   changes that configuration value when a dialog is running and checking
   its value.