Added SDP I header. It describes a file transfer.
Added a MSRP (Message Session Relay Protocol) implementation.

Added SIP state machines to set up a file transfer session.

Added a method to SdpHeaderA to get filename/size/hash/type from the file
selector attribute in the A header.

The above code is not enabled unless libminisip is configured 
with "--enable-msrp". (this first commit will probably fail if msrp is not
enabled).