libmutil:  * Added FileSystem.{h,cxx} to the MSVC project. libmnetutil:  * Made the HttpDownloader use references to sockets instead of pointers.  * Created a HttpFileSystem implementation. In this commit no write support    is enabled.
Minisip reads (and sometimes writes) several files when it is
executed:
 - configuration file (if not using gconf)
 - certificates
 - address books
 - ring tones and other UI resources
 (- plugins)
 
We have explored the possibility to store settins online (see "Secure on-line
configuration for SIP User Agents" on the publications section of
minisip.org), and there are both the kind of files that you always would
like to store on the device (information such as what is the sound i/o to
use on this device), and information that some users would always like to
store online.

I believe that we should create two "FileSystem" objects,
 - one that is device-local ones, and
 - one that can be local, but also on-line.

If data is retrieved using HTTP+TLS_SRP, or opened locally will be transparent 
to any code after the file systems have been initialized.

Example of code using FileSystem objects:

  void process(string fileurl){
    MRef<FileSystem*> fs;
    if ( path.find("http://") == 0 ){
      fs = new HttpFileSystem( new TCPSocket("www.minisip.org", 80) );
    }else{
      fs = new LocalFileSystem;
    }

    //Filesystem independent I/O from here
    MRef<File*> f = fs->open(argv[1]);
    cerr << "file size: "<< f->size() << endl;

    f->read( buf, len );
  }